+++ packages/odata-service-inquirer/src/prompts/datasources/sap-system/abap-on-prem/questions.ts
@@ -100,11 +100,7 @@ export function getAbapOnPremSystemQuestions(
                 return valRes;
             }
         } as InputQuestion<AbapOnPremAnswers>,
-        ...getCredentialsPrompts<AbabpOnPremCredentialsAnswers>(
-            connectValidator,
-            abapOnPremPromptNamespace,
-            sapClientRef
-        )
+        ...getCredentialsPrompts<AbapOnPremAnswers>(connectValidator, abapOnPremPromptNamespace, sapClientRef)
     ];
 
     if (systemNamePromptOptions?.hide !== true) {