+++ packages/fiori-elements-writer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/fiori-elements-writer
 
+## 2.4.11
+
+### Patch Changes
+
+-   Updated dependencies [15ec5c4]
+    -   @sap-ux/fiori-generator-shared@0.12.5
+    -   @sap-ux/cap-config-writer@0.10.5
+
 ## 2.4.10
 
 ### Patch Changes