+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 0.1.6
+
+### Patch Changes
+
+-   Updated dependencies [15ec5c4]
+    -   @sap-ux/fiori-generator-shared@0.12.5
+    -   @sap-ux/ui5-library-inquirer@0.4.6
+
 ## 0.1.5
 
 ### Patch Changes